Genericizes `ConditionCache` into `GraphCache` and uses it within `allSwiftPluginExecutables`.

### Short description 📝

This is a quick caching implementation in `allSwiftPluginExecutables` which should help with the performance issues identified in [this post](https://community.tuist.io/t/improve-tuists-performance/130/2)

I wanted to toss this up as an option for perf improvements but may not have the personal time to see it landed.  Open to comments about needed tests or other changes but totally fine for someone else to take this the rest of the way (and actually do things like "measure improvements" and "verify it's not actually a regression").

I'm fairly confident in this change tho since it's the same structured we used for platform conditions.

### How to test the changes locally 🧐

Everything should keep working.

### Contributor checklist ✅

- [x] The code has been linted using run `mise run lint-fix`
- [ ] The change is tested via unit testing or acceptance testing, or both
- [ ] The title of the PR is formulated in a way that is usable as a changelog entry
- [x] In case the PR introduces changes that affect users, the documentation has been updated

### Reviewer checklist ✅

- [ ] The code architecture and patterns are consistent with the rest of the codebase
- [ ] Reviewer has checked that, if needed, the PR includes the label `changelog:added`, `changelog:fixed`, or `changelog:changed`, and the title is usable as a changelog entry