@fortmarek 2024-10-29
Hey @waltflanagan 👋 

This does improve the performance a ton – before, the method took around ~ 5 % of the CPU time, now it's basically zero (0.1).

Benchmarking a real-world project leads to around ~7 % performance improvements of the overall generation – which is huge considering this was just one method 🤯 

```
## Before

Benchmark 1: "/Users/marekfort/Library/Developer/Xcode/DerivedData/TuistExt-fnvyzgrgowotpwgcllhszvqbqzxi/Build/Products/Release/tuist_ext" generate --no-open
  Time (mean ± σ):      4.503 s ±  0.132 s    [User: 9.485 s, System: 3.667 s]
  Range (min … max):    4.388 s …  4.827 s    10 runs

## After
Benchmark 1: "/Users/marekfort/Library/Developer/Xcode/DerivedData/TuistExt-fnvyzgrgowotpwgcllhszvqbqzxi/Build/Products/Release/tuist_ext" generate --no-open
  Time (mean ± σ):      4.177 s ±  0.088 s    [User: 9.421 s, System: 3.779 s]
  Range (min … max):    4.075 s …  4.360 s    10 runs
```

I'll do a proper review tomorrow, possibly updating things myself if needed 🙂