@waltflanagan 2024-10-29
amazing! thanks @fortmarek.  let me know if you need anything else.